# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MemoryActivityResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "memory_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Id"
          },
          "operation": {
            "enum": [
              "read",
              "create",
              "update",
              "delete"
            ],
            "title": "Operation",
            "type": "string"
          },
          "result_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Count"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "operation",
          "memory_id",
          "content",
          "result_count",
          "task_id",
          "created_at"
        ],
        "title": "MemoryActivityResponse",
        "type": "object"
      },
      "MemoryDreamDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/MemoryActivityResponse"
            },
            "title": "Activities",
            "type": "array"
          },
          "activity_total": {
            "title": "Activity Total",
            "type": "integer"
          },
          "dream": {
            "$ref": "#/components/schemas/MemoryDreamResponse"
          }
        },
        "required": [
          "dream",
          "activities",
          "activity_total"
        ],
        "title": "MemoryDreamDetailResponse",
        "type": "object"
      },
      "MemoryDreamListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryDreamResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "latest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryDreamResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "latest"
        ],
        "title": "MemoryDreamListResponse",
        "type": "object"
      },
      "MemoryDreamResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "target_date",
          "status",
          "result",
          "failure_reason",
          "created_at",
          "started_at",
          "completed_at"
        ],
        "title": "MemoryDreamResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AniuSessionCookie": {
        "in": "cookie",
        "name": "aniu_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "基于股票交易的可进化智能体系统",
    "title": "Aniu",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/aniu/memory-dreams": {
      "get": {
        "operationId": "list_memory_dreams_api_aniu_memory_dreams_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDreamListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "List Memory Dreams",
        "tags": [
          "MemoryDreams"
        ]
      }
    },
    "/api/aniu/memory-dreams/run": {
      "post": {
        "operationId": "run_memory_dream_api_aniu_memory_dreams_run_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDreamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Run Memory Dream",
        "tags": [
          "MemoryDreams"
        ]
      }
    },
    "/api/aniu/memory-dreams/{task_id}": {
      "delete": {
        "operationId": "delete_memory_dream_api_aniu_memory_dreams__task_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Delete Memory Dream",
        "tags": [
          "MemoryDreams"
        ]
      },
      "get": {
        "operationId": "get_memory_dream_api_aniu_memory_dreams__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDreamDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Memory Dream",
        "tags": [
          "MemoryDreams"
        ]
      }
    },
...
  }
}